refactor: potential fix for flaky async client notify tests
uses a countdown latch rather than sleeping for 1ms to check whether a method is async. this seems
to fail rather consistently at the moment on travis API 21 emulators, which are faster than the
16/19
